requirements.txt
flask
numpy
pandas
flask
flask_mysqldb
mysql-connector-python